add new tool window for functions
