+++ src/test/java/com/redhat/devtools/intellij/knative/BaseTest.java
@@ -22,7 +22,7 @@ import com.redhat.devtools.intellij.knative.tree.KnEventingTriggersNode;
 import com.redhat.devtools.intellij.knative.tree.KnFunctionLocalNode;
 import com.redhat.devtools.intellij.knative.tree.KnFunctionNode;
 import com.redhat.devtools.intellij.knative.tree.KnFunctionsNode;
-import com.redhat.devtools.intellij.knative.tree.KnFunctionsTreeStructure;
+import com.redhat.devtools.intellij.knative.tree.KnLocalFunctionsTreeStructure;
 import com.redhat.devtools.intellij.knative.tree.KnRevisionNode;
 import com.redhat.devtools.intellij.knative.tree.KnRootNode;
 import com.redhat.devtools.intellij.knative.tree.KnServiceNode;
@@ -55,7 +55,7 @@ public class BaseTest {
     protected KnFunctionNode knFunctionNode;
     protected KnFunctionLocalNode knFunctionLocalNode;
     protected KnTreeStructure knTreeStructure;
-    protected KnFunctionsTreeStructure knFunctionsTreeStructure;
+    protected KnLocalFunctionsTreeStructure knFunctionsTreeStructure;
     protected KnEventingBrokerNode knEventingBrokerNode;
     protected KnEventingChannelsNode knEventingChannelsNode;
     protected KnEventingSourcesNode knEventingSourcesNode;
@@ -75,7 +75,7 @@ public class BaseTest {
         knEventingNode = mock(KnEventingNode.class);
         knRevisionNode = mock(KnRevisionNode.class);
         knTreeStructure = mock(KnTreeStructure.class);
-        knFunctionsTreeStructure = mock(KnFunctionsTreeStructure.class);
+        knFunctionsTreeStructure = mock(KnLocalFunctionsTreeStructure.class);
         knEventingBrokerNode = mock(KnEventingBrokerNode.class);
         knEventingChannelsNode = mock(KnEventingChannelsNode.class);
         knEventingSourcesNode = mock(KnEventingSourcesNode.class);
