+++ src/test/java/com/redhat/devtools/intellij/knative/utils/TreeHelperTest.java
@@ -20,7 +20,7 @@ import com.intellij.ui.treeStructure.Tree;
 import com.redhat.devtools.intellij.knative.Constants;
 import com.redhat.devtools.intellij.knative.FixtureBaseTest;
 import com.redhat.devtools.intellij.knative.kn.Kn;
-import com.redhat.devtools.intellij.knative.tree.KnFunctionsTreeStructure;
+import com.redhat.devtools.intellij.knative.tree.KnLocalFunctionsTreeStructure;
 import com.redhat.devtools.intellij.knative.tree.KnTreeStructure;
 import javax.swing.JViewport;
 import org.apache.commons.lang.StringUtils;
@@ -30,7 +30,7 @@ import org.junit.Test;
 import org.mockito.MockedStatic;
 
 
-import static com.redhat.devtools.intellij.knative.Constants.KNATIVE_FUNC_TOOL_WINDOW_ID;
+import static com.redhat.devtools.intellij.knative.Constants.KNATIVE_LOCAL_FUNC_TOOL_WINDOW_ID;
 import static com.redhat.devtools.intellij.knative.Constants.KNATIVE_TOOL_WINDOW_ID;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
@@ -201,21 +201,21 @@ public class TreeHelperTest extends FixtureBaseTest {
     }
 
     @Test
-    public void GetKnFunctionsTreeStructure_ProjectIsNull_Null() {
-        assertNull(TreeHelper.getKnFunctionsTreeStructure(null));
+    public void GetLocalKnFunctionsTreeStructure_ProjectIsNull_Null() {
+        assertNull(TreeHelper.getKnLocalFunctionsTreeStructure(null));
     }
 
     @Test
-    public void GetKnFunctionsTreeStructure_ProjectWithoutClientProperty_Null() {
-        getTree(KNATIVE_FUNC_TOOL_WINDOW_ID);
+    public void GetLocalKnFunctionsTreeStructure_ProjectWithoutClientProperty_Null() {
+        getTree(KNATIVE_LOCAL_FUNC_TOOL_WINDOW_ID);
         when(tree.getClientProperty(Constants.STRUCTURE_PROPERTY)).thenReturn(null);
-        assertNull(TreeHelper.getKnFunctionsTreeStructure(project));
+        assertNull(TreeHelper.getKnLocalFunctionsTreeStructure(project));
     }
 
     @Test
-    public void GetKnFunctionsTreeStructure_Project_KnTreeStructure() {
-        getKnFunctionsTreeStructure();
-        KnFunctionsTreeStructure resultingStructure = TreeHelper.getKnFunctionsTreeStructure(project);
+    public void GetLocalKnFunctionsTreeStructure_Project_KnTreeStructure() {
+        getKnLocalFunctionsTreeStructure();
+        KnLocalFunctionsTreeStructure resultingStructure = TreeHelper.getKnLocalFunctionsTreeStructure(project);
         assertNotNull(resultingStructure);
         assertEquals(knFunctionsTreeStructure, resultingStructure);
     }
@@ -271,7 +271,7 @@ public class TreeHelperTest extends FixtureBaseTest {
 
     @Test
     public void RefreshFunc_ProjectWithoutTreeStructure_Nothing() {
-        getTree(KNATIVE_FUNC_TOOL_WINDOW_ID);
+        getTree(KNATIVE_LOCAL_FUNC_TOOL_WINDOW_ID);
         when(tree.getClientProperty(Constants.STRUCTURE_PROPERTY)).thenReturn(null);
         TreeHelper.refreshFunc(project);
         verify(knFunctionsTreeStructure, never()).fireModified(any());
@@ -279,7 +279,7 @@ public class TreeHelperTest extends FixtureBaseTest {
 
     @Test
     public void RefreshFunc_ProjectAndNodeAreValid_Refresh() {
-        getKnFunctionsTreeStructure();
+        getKnLocalFunctionsTreeStructure();
         TreeHelper.refreshFunc(project);
         verify(knFunctionsTreeStructure).fireModified(any());
     }
@@ -299,8 +299,8 @@ public class TreeHelperTest extends FixtureBaseTest {
         when(tree.getClientProperty(Constants.STRUCTURE_PROPERTY)).thenReturn(knTreeStructure);
     }
 
-    private void getKnFunctionsTreeStructure() {
-        getTree(KNATIVE_FUNC_TOOL_WINDOW_ID);
+    private void getKnLocalFunctionsTreeStructure() {
+        getTree(KNATIVE_LOCAL_FUNC_TOOL_WINDOW_ID);
         when(tree.getClientProperty(Constants.STRUCTURE_PROPERTY)).thenReturn(knFunctionsTreeStructure);
     }
 }
